Completed tests and fixxed duplicate error.
