+++ api/src/paths/user/add.ts
@@ -1,12 +1,12 @@
 import { RequestHandler } from 'express';
 import { Operation } from 'express-openapi';
-import { SYSTEM_ROLE } from '../constants/roles';
-import { getDBConnection } from '../database/db';
-import { HTTP400 } from '../errors/CustomError';
-import { addUserSystemRoles } from './user/{userId}/system-roles/update';
-import { addSystemUser, getSystemUser } from '../paths-helpers/system-user';
-import { authorizeRequestHandler } from '../request-handlers/security/authorization';
-import { getLogger } from '../utils/logger';
+import { SYSTEM_ROLE } from '../../constants/roles';
+import { getDBConnection } from '../../database/db';
+import { HTTP400 } from '../../errors/CustomError';
+import { addUserSystemRoles } from './{userId}/system-roles/update';
+import { ensureSystemUser, getSystemUser } from '../../paths-helpers/system-user';
+import { authorizeRequestHandler } from '../../request-handlers/security/authorization';
+import { getLogger } from '../../utils/logger';
 
 const defaultLog = getLogger('paths/user');
 
@@ -107,7 +107,7 @@ export function addSystemRoleUser(): RequestHandler {
     try {
       await connection.open();
 
-      await addSystemUser(userIdentifier, identitySource, connection);
+      await ensureSystemUser(userIdentifier, identitySource, connection);
 
       await connection.commit();
